Refactor v2 trace arg helpers
